+++ lib/ypsiloid/src/core/ypsiloid_core.scala
@@ -61,7 +61,7 @@ extension (yaml: Yaml.Ast)
   inline def isNull:    Boolean = yaml.asInstanceOf[AnyRef | Null] == null
   inline def isLong:    Boolean = yaml.isInstanceOf[Long]
   inline def isDouble:  Boolean = yaml.isInstanceOf[Double]
-  inline def isBcd:     Boolean = yaml.isInstanceOf[Array[Long]]
+  inline def isBcd:     Boolean = yaml.isInstanceOf[Array[Double]]
   inline def isNumber:  Boolean = isLong || isDouble || isBcd
   inline def isString:  Boolean = yaml.isInstanceOf[String]
   inline def isBoolean: Boolean = yaml.isInstanceOf[Boolean]
@@ -119,17 +119,17 @@ extension (yaml: Yaml.Ast)
   def double(using Tactic[YamlError]): Double = yaml.asInstanceOf[Matchable] match
     case value: Double                 => value
     case value: Long                   => value.toDouble
-    case value: Array[Long] @unchecked => value.asInstanceOf[jacinta.Bcd].toDouble
+    case value: Array[Double] @unchecked => value.asInstanceOf[jacinta.Bcd].toDouble
     case _                             => expected(YamlPrimitive.Decimal) yet 0.0
 
   def long(using Tactic[YamlError]): Long = yaml.asInstanceOf[Matchable] match
     case value: Long                   => value
     case value: Double                 => value.toLong
-    case value: Array[Long] @unchecked => value.asInstanceOf[jacinta.Bcd].toLong.or(0L)
+    case value: Array[Double] @unchecked => value.asInstanceOf[jacinta.Bcd].toLong.or(0L)
     case _                             => expected(YamlPrimitive.Integer) yet 0L
 
   def bcd(using Tactic[YamlError]): jacinta.Bcd = yaml.asInstanceOf[Matchable] match
-    case value: Array[Long] @unchecked => value.asInstanceOf[jacinta.Bcd]
+    case value: Array[Double] @unchecked => value.asInstanceOf[jacinta.Bcd]
     case value: Long                   => jacinta.Bcd(BigDecimal(value))
     case value: Double                 => jacinta.Bcd(BigDecimal(value))
